+++ signal_process/ecg_ht.cpp
@@ -7,19 +7,19 @@
 using namespace std;
 
 //Differentiation stage 
-vector<double> ht_differentiation(const vector<double>& phys_ecg_data, int sampling_freq){
-    vector<double> differentiated_ecg;
+vector<float> ht_differentiation(const vector<float>& phys_ecg_data, int sampling_freq){
+    vector<float> differentiated_ecg;
     //differentiate the ecg data using the difference equation
     for(int i = 1; i < phys_ecg_data.size(); i++){
-        double derivative = (phys_ecg_data[i] - phys_ecg_data[i-1]) * sampling_freq;
+        float derivative = (phys_ecg_data[i] - phys_ecg_data[i-1]) * sampling_freq;
         differentiated_ecg.push_back(derivative);
     }
     return differentiated_ecg;
 }
 
 //Squaring stage 
-vector<double> ht_squaring(const vector<double>& differentiated_ecg){
-    vector<double> squared_ecg;
+vector<float> ht_squaring(const vector<float>& differentiated_ecg){
+    vector<float> squared_ecg;
     
     for(int i = 0; i < differentiated_ecg.size(); i++){
         squared_ecg.push_back(differentiated_ecg[i] * differentiated_ecg[i]);  
@@ -28,9 +28,9 @@ vector<double> ht_squaring(const vector<double>& differentiated_ecg){
 }
 
 //Moving Average stage 
-vector<double> ht_moving_average(const vector<double>& squared_ecg, int window_size){
-    double sum = 0.0;
-    vector<double> averaged_ecg;
+vector<float> ht_moving_average(const vector<float>& squared_ecg, int window_size){
+    float sum = 0.0;
+    vector<float> averaged_ecg;
     
     for(int j = 0; j < squared_ecg.size(); j++){
         // If window is full, remove the oldest sample (sliding window)
@@ -47,18 +47,18 @@ vector<double> ht_moving_average(const vector<double>& squared_ecg, int window_s
 }
 
 //Adaptive Threshold stage 
-vector<double> ht_adaptive_threshold(const vector<double>& averaged_ecg, double threshold_factor, int sampling_rate) {
-    double SPKI, NPKI;
-    double peak_threshold, noise_threshold;
-    vector<double> approx_indices;
+vector<float> ht_adaptive_threshold(const vector<float>& averaged_ecg, float threshold_factor, int sampling_rate) {
+    float SPKI, NPKI;
+    float peak_threshold, noise_threshold;
+    vector<float> approx_indices;
     // Initialize
     /*int init_segment_length = max(1, (int)(averaged_ecg.size() * 0.05));
-    double PEAKI = *max_element(averaged_ecg.begin(), averaged_ecg.begin() + init_segment_length);
+    float PEAKI = *max_element(averaged_ecg.begin(), averaged_ecg.begin() + init_segment_length);
     */
 
     int start_index = 70; // skip samples
     int init_segment_length = max(1, (int)((averaged_ecg.size() - start_index) * 0.05));
-    double PEAKI = *max_element(averaged_ecg.begin() + start_index, averaged_ecg.begin() + start_index + init_segment_length);
+    float PEAKI = *max_element(averaged_ecg.begin() + start_index, averaged_ecg.begin() + start_index + init_segment_length);
 
     SPKI = 0.125 * PEAKI;
     NPKI = 0.125 * PEAKI;
@@ -70,7 +70,7 @@ vector<double> ht_adaptive_threshold(const vector<double>& averaged_ecg, double 
     int last_r_index = -minimum_samples;
 
     for (int i = start_index; i < averaged_ecg.size() - 1; i++) { 
-        double current_val = averaged_ecg[i];
+        float current_val = averaged_ecg[i];
 
         // Simple local max check
         if (current_val > averaged_ecg[i - 1] && current_val > averaged_ecg[i + 1]) {
@@ -93,8 +93,8 @@ vector<double> ht_adaptive_threshold(const vector<double>& averaged_ecg, double 
 }
 
 //R Peak detection stage 
-vector<double> ht_peak_detection(const vector<double>& threshold_ecg, const vector<double>& filtered_ecg, int window_size){
-    vector<double> r_peak_indices;
+vector<float> ht_peak_detection(const vector<float>& threshold_ecg, const vector<float>& filtered_ecg, int window_size){
+    vector<float> r_peak_indices;
     
     for(int i = 0; i < threshold_ecg.size(); i++){
         int approx_index = (int)threshold_ecg[i];
@@ -102,10 +102,10 @@ vector<double> ht_peak_detection(const vector<double>& threshold_ecg, const vect
         int end = min((int)filtered_ecg.size() - 1, approx_index + window_size);
 
         int best_index = approx_index;
-        double best_val = filtered_ecg[approx_index];
+        float best_val = filtered_ecg[approx_index];
 
         for(int j = start; j <= end; j++){
-            double val = filtered_ecg[j];
+            float val = filtered_ecg[j];
             if(val > best_val){
                 best_val = val;
                 best_index = j;
